examples/flask_app/requirements.txt
Flask==2.3.2
SQLAlchemy==1.4.39
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
Werkzeug==2.3.6
